redmineID #5845 - Updated ivy config to include keystore.
The keystore is now in nexus, as com.adaptris:code-signing. It lives
in the adaptris-private-releases repository.
Added in ivy configuration to include the keystore as a specific
configuration and downloads it as part of the "prepare" phase.

Modified ant-ivy-includes so that signing now takes place when jar
creation occurs; can be turned off by explicitly setting the property
unsigned.code=true in build.properties.