INTERLOK-293 Standardised the manifest file. INTERLOK-293 #time 1h 30m
Added in a macrodef that contains a standard manifest file which includes
the Implementation-Version, but no additional sections.
Made all sub-components call adapter:create-manifest when where they
were creating a manifest file.

The sections were being mixed up anyway because of the signing aspect
and makes no sense now that we have pretty much a single jar for a
single component.